Implement feature to map compatible props of betwen two object of different CLR types.
With is this push, the core features and proof-of-concept
as it pertains to object mapping have been done for the  library.

That is a good start for a single weekend, or half-weekend.

More features will be implemented from the backlog.

At a high level the behaviors implemented include:
 - void Map<T>(T source, T target)
 - void Map(object source, object target)
 - void MapFrom<T>(T source, T target)
 - void MapFrom(object source, object target)
 - void MapTo<T>(T source, T target)
 - void MapTo(object source, object target)

As always, I say:
"No problem can withstand the assault of sustained thinking"
- Voltaire.

This is Kingsley signing off for now.