Switched the license from MIT to GPL v3
I decided to switch the license to a more copy-left variant. GNU GENERAL PUBLIC LICENSE, Version 3, 29 June 2007. Also known as the GPL v3.

After racking my brain to solve complex problems and encapsulating them in a library, I think it is fair to say that if people want to use my work down the road, I think it is not out of place to ask that they do so in a way that is fair.

I am pretty sure that is what they would want for themselves also.

Cheers.

Kingsley Omon-Edo