Proof-of-Concept of mapping mapping to new instances. Done. No regressions.
Using the .NET Reflections API, by making a call to Activator.CreateInstance<T>(), I was able to map from from a given source type to a newly instantiated target type.

This significantly enhances the sophistication of the mapping library as it obviates the need to supply pre-canned objects manually for mapping.

Will have to review / augment the signatures of the existing methods for both single object and collection mapping.

As always,

"There is no problem that can withstand the assault of sustained thinking"
- Voltaire.

Though this was a piece of cake.
- Kingsley Omon-Edo.